module_classify/requirements.txt
accelerate==1.1.1
aenum==3.1.15
affine==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.16
albumentations==1.4.10
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
asttokens==2.4.1
attrs==24.2.0
bitsandbytes==0.44.1
blinker==1.9.0
Bottleneck==1.4.2
cachetools==5.5.0
certifi==2024.8.30
cftime==1.6.4.post1
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
dask==2024.11.1
databricks-sdk==0.36.0
decorator==5.1.1
Deprecated==1.2.14
docker==7.1.0
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.8.0
executing==2.1.0
fasteners==0.19
filelock==3.16.1
fiona==1.10.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
granitewxc @ git+https://github.com/IBM/granite-wxc.git@1d180f6d730e26fcd605925f6514f7974122f973
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h5netcdf==1.4.0
h5py==3.12.1
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==4.34.0
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
lightly==1.5.14
lightly-utils==0.0.2
lightning==2.4.0
lightning-utilities==0.11.8
locket==1.0.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
netCDF4==1.7.2
networkx==3.4.2
numcodecs==0.13.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
peft==0.12.0
pexpect==4.9.0
pillow==11.0.0
pretrainedmodels==0.7.4
PrithviWxC @ git+https://github.com/NASA-IMPACT/Prithvi-WxC.git@ecfd69b2e94b6505d3accbdecf095b92fd18411e
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
rasterio==1.3.11
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rioxarray==0.18.1
rsa==4.9
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
segmentation-models-pytorch==0.3.4
shapely==2.0.6
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
tensorboardX==2.6.2.2
-e git+ssh://git@github.com/IBM/terratorch.git@ca289f0663a84952247213958cf3727a07359a87#egg=terratorch
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.7
tokenizers==0.20.3
toolz==1.0.0
torch==2.5.0
torcheval==0.0.7
torchgeo==0.6.1
torchmetrics==1.3.1
torchvision==0.20.0
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wget==3.2
wrapt==1.16.0
xarray==2024.10.0
yacs==0.1.8
yarl==1.17.1
zarr==2.18.3
zipp==3.21.0
python-dotenv
